ci: refactor script, add fixable icons
